Latest document for macOS Users
The previous document for installing Jupyter Notebook for macOS users caused errors for some users. This latest document is much simplified. This contains steps of how the user can install all the necessary tools in their macOS by running a shell script in the bash terminal with very few commands